+++ tests/Unit/Public/ConvertTo-DifferenceString.tests.ps1
@@ -155,4 +155,13 @@ Describe '-join (ConvertTo-DifferenceString' {
         $result | Should -Match "`e\[31m0D`e\[0m `e\[31m0A`e\[0m"  # Carriage return + Newline
         $result | Should -Match "`e\[31m0A`e\[0m `e\[31m0D`e\[0m"  # Newline + Carriage return
     }
+
+    # This could not be used in Azure DevOps pipeline due to irregular performance in build agents.
+    # It 'Should process large strings efficiently' {
+    #     $largeString1 = 'a' * 10000
+    #     $largeString2 = 'a' * 9999 + 'b'
+
+    #     # TODO: When Because works in Should-BeFasterThan, uncomment the part on the following line.
+    #     Measure-Command { ConvertTo-DifferenceString -ReferenceString $largeString1 -DifferenceString $largeString2 } | Should-BeFasterThan '1.5s' #-Because 'Large strings should be processed efficiently, and should not take more than 1.5 seconds.'
+    # }
 }